bmlife/requirements.txt
certifi==2019.3.9
Django==2.1.8
django-filter==2.1.0
django-debug-toolbar==1.11
django-extensions==2.1.6
django-model-utils==3.1.2
djangorestframework==3.9.4
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
Markdown==3.1.1
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
